If .zip not found, look in the /old/ folder
Modified the gutenberg_download(...) function to look in the /old/ folder on http://aleph.gutenberg.org/.../old/ in case the .zip file is not in the normal file folder.

I had an issue when working through Text Mining with R by Julia Silge & David Robinson. In Chapter 3, "A Corpus of Physics Texts" gutenberg ID 5001 (Albert Einstein's The Special and General Theory) .zip file was not where it was supposed to be. I went to the URL and the .zip was in a directory labeled /old/. Adding my changes to the gutenberg_download(...) function tells R to look in the /old/ folder if it can't find the .zip in the regular folder.